feat(db): migrate from MySQL to PostgreSQL
- Update sqlc to use pgx/v5 driver
- Convert SQL queries to PostgreSQL syntax
- Change ID types from uint32 to int64 (bigint)
- Replace sql.Null* with pgtype equivalents
- Update all packages for *pgxpool.Pool connections

+++ scorer/types/types.go
@@ -9,6 +9,6 @@ import (
 
 type Scorer interface {
 	// Lookback(LookbackOptions)
-	Setup(id uint32)
+	Setup(id int64)
 	Score(ctx context.Context, db *ntpdb.Queries, serverScore ntpdb.ServerScore, ls ntpdb.LogScore) (score.Score, error)
 }
